Create basic main.cpp and Makefile
This creates a basic main.cpp and Makefile. The source file compiles
to a program that takes a single argument (intended to be a filename)
and simply prints it out. It raises a usage message on any other
invocation.

The Makefile sets the C++ version used to be C++20, as we may make
use of features such as concepts & constraints, lambda templates,
and <filesystem> during the development of the project.

NOTE: The compilation files exclude UBSan and ASan on Windows, as
MinGW and MSYS do not provide the relevant libraries in the default
installation.